+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 -   Adjust task-item css class to prevent css conflicts. #7593
+-   Update task-item logic to only display content when expanded is true. #7611
 
 # 1.5.1